Test object to table in LaTeX format

Transform a test object to table in LaTeX format

# S3 method for PosteriorDirichlet
htest2Tex(test)

Arguments

test
Test object with pvalue(s), test name and statistic(s)

Value

This method prints the necessary code for include a table with the information provided by the test.

Examples

htest2Tex(cd.test(results))

#> [1] "\\begin{table}[] \n\\centering\n\\caption{Chakraborti and Desu test} \n\\begin{tabular}{lll} \n\\hline\n\\multicolumn{3}{c}{Chakraborti and Desu} \\\\ \\hline\n\n\\multirow{1}{*}{data.name}\n\t & \t\t & \tresults\t \\\\ \\hline\n\\multirow{2}{*}{statistic}\n\t & \tmedian\t & \t 0.8336\t\\\\\n \t & \tW\t & \t65.0000\t \\\\ \\hline\n\\multirow{2}{*}{p.value}\n\t & \tExact p-value\t & \t8.308e-06\t\\\\\n \t & \tAsymtotic p-value\t & \t1.000e+00\t \\\\ \\hline\n\\end{tabular}\n\\end{table}"