Number of runs up and down median test for randomness

This function performs the Number of runs up and down median test

numberRunsUpDownMedian.test(sequence)

Arguments

sequence
Sequence of data

Value

A list with pvalues for alternative hypothesis, statistics, method and data name

Examples

 numberRunsUpDownMedian.test(c(23,99,89,36,1,64,14,28,2,81))

#> $data.name
#> [1] "c(23, 99, 89, 36, 1, 64, 14, 28, 2, 81)"
#> 
#> $statistic
#> Number of elements     Number of runs             Median 
#>                 10                  6                 32 
#> 
#> $p.value
#>        Exact Left Tail       Exact Right Tail      Exact Double Tail 
#>              0.6430000              0.6430000              1.0000000 
#>   Asymptotic Left Tail  Asymptotic Right Tail Asymptotic Double Tail 
#>              0.6313422              0.6313422              1.0000000 
#> 
#> $method
#> [1] "Runs Up and Down With respect of the median"
#>